Moved IT tests and reconfigured CI
