+++ pkg/bindings/kafka_sarama/producer_message_encoder_test.go
@@ -1,5 +1,3 @@
-// +build kafka
-
 package kafka_sarama
 
 import (
