+++ test/integration/amqp_binding/amqp_test.go
@@ -1,6 +1,4 @@
-// +build amqp
-
-package amqp
+package amqp_binding
 
 import (
 	"context"
@@ -15,6 +13,7 @@ import (
 
 	"github.com/cloudevents/sdk-go/pkg/binding"
 	"github.com/cloudevents/sdk-go/pkg/binding/test"
+	amqp2 "github.com/cloudevents/sdk-go/pkg/bindings/amqp"
 	"github.com/cloudevents/sdk-go/pkg/event"
 )
 
@@ -99,13 +98,13 @@ func testClient(t testing.TB) (client *amqp.Client, session *amqp.Session, addr 
 	return client, session, addr
 }
 
-func testSenderReceiver(t testing.TB, senderOptions ...SenderOptionFunc) (io.Closer, binding.Sender, binding.Receiver) {
+func testSenderReceiver(t testing.TB, senderOptions ...amqp2.SenderOptionFunc) (io.Closer, binding.Sender, binding.Receiver) {
 	c, ss, a := testClient(t)
 	r, err := ss.NewReceiver(amqp.LinkSourceAddress(a))
 	require.NoError(t, err)
 	s, err := ss.NewSender(amqp.LinkTargetAddress(a))
 	require.NoError(t, err)
-	return c, NewSender(s, senderOptions...), &receiver{r}
+	return c, amqp2.NewSender(s, senderOptions...), &amqp2.Receiver{r}
 }
 
 func BenchmarkSendReceive(b *testing.B) {
