+++ test/integration/kafka_sarama_binding/kafka_test.go
@@ -1,6 +1,4 @@
-// +build kafka
-
-package kafka_sarama
+package kafka_sarama_binding
 
 import (
 	"context"
@@ -16,59 +14,12 @@ import (
 	cloudevents "github.com/cloudevents/sdk-go"
 	"github.com/cloudevents/sdk-go/pkg/binding"
 	"github.com/cloudevents/sdk-go/pkg/binding/test"
+	"github.com/cloudevents/sdk-go/pkg/bindings/kafka_sarama"
 	"github.com/cloudevents/sdk-go/pkg/event"
 )
 
 const (
 	TEST_GROUP_ID = "test_group_id"
-)
-
-var (
-	e                                   = test.FullEvent()
-	StructuredConsumerMessageWithoutKey = &sarama.ConsumerMessage{
-		Value: test.MustJSON(e),
-		Headers: []*sarama.RecordHeader{{
-			Key:   []byte(contentTypeHeader),
-			Value: []byte(cloudevents.ApplicationCloudEventsJSON),
-		}},
-	}
-	StructuredConsumerMessageWithKey = &sarama.ConsumerMessage{
-		Key:   []byte("aaa"),
-		Value: test.MustJSON(e),
-		Headers: []*sarama.RecordHeader{{
-			Key:   []byte(contentTypeHeader),
-			Value: []byte(cloudevents.ApplicationCloudEventsJSON),
-		}},
-	}
-	BinaryConsumerMessageWithoutKey = &sarama.ConsumerMessage{
-		Value: []byte("hello world!"),
-		Headers: mustToSaramaConsumerHeaders(map[string]string{
-			"ce_type":            e.Type(),
-			"ce_source":          e.Source(),
-			"ce_id":              e.ID(),
-			"ce_time":            test.Timestamp.String(),
-			"ce_specversion":     "1.0",
-			"ce_dataschema":      test.Schema.String(),
-			"ce_datacontenttype": "text/json",
-			"ce_subject":         "topic",
-			"ce_exta":            "someext",
-		}),
-	}
-	BinaryConsumerMessageWithKey = &sarama.ConsumerMessage{
-		Key:   []byte("akey"),
-		Value: []byte("hello world!"),
-		Headers: mustToSaramaConsumerHeaders(map[string]string{
-			"ce_type":            e.Type(),
-			"ce_source":          e.Source(),
-			"ce_id":              e.ID(),
-			"ce_time":            test.Timestamp.String(),
-			"ce_specversion":     "1.0",
-			"ce_dataschema":      test.Schema.String(),
-			"ce_datacontenttype": "text/json",
-			"ce_subject":         "topic",
-			"ce_exta":            "someext",
-		}),
-	}
 )
 
 func TestSendStructuredMessageToStructuredWithKey(t *testing.T) {
@@ -153,12 +104,12 @@ func testClient(t testing.TB) sarama.Client {
 	return client
 }
 
-func testSenderReceiver(t testing.TB, options ...SenderOptionFunc) (func(), binding.Sender, binding.Receiver) {
+func testSenderReceiver(t testing.TB, options ...kafka_sarama.SenderOptionFunc) (func(), binding.Sender, binding.Receiver) {
 	client := testClient(t)
 
 	topicName := "test-ce-client-" + uuid.New().String()
-	r := NewReceiver(client, TEST_GROUP_ID, topicName)
-	s, err := NewSender(client, topicName, options...)
+	r := kafka_sarama.NewReceiver(client, TEST_GROUP_ID, topicName)
+	s, err := kafka_sarama.NewSender(client, topicName, options...)
 	require.NoError(t, err)
 
 	return func() {
@@ -175,14 +126,4 @@ func BenchmarkSendReceive(b *testing.B) {
 	c, s, r := testSenderReceiver(b)
 	defer c() // Cleanup
 	test.BenchmarkSendReceive(b, s, r)
-}
-
-func mustToSaramaConsumerHeaders(m map[string]string) []*sarama.RecordHeader {
-	res := make([]*sarama.RecordHeader, len(m))
-	i := 0
-	for k, v := range m {
-		res[i] = &sarama.RecordHeader{Key: []byte(k), Value: []byte(v)}
-		i++
-	}
-	return res
 }
